image insert Done and delete alert add

+++ crud/formView.php
@@ -59,7 +59,7 @@
                 $userName =  $rows['userName'] ;
                 $email =  $rows['email'] ;
                 $password =  $rows['password'] ;
-                
+
                 $insImageadd = $rows['profile_photo'];
 
             ?>
@@ -70,7 +70,7 @@
                         <td><img width='50px' src="image/<?php echo $insImageadd?>" alt="" srcset=""></td>
                         <td><?php echo $email;?></td>
                         <td><?php echo $password;?></td>
-                        <td><a href="single_data_edit.php?edit_Id=<?php echo $dbId ?>;">Edit</a> || <a href="delete.php?id=<?php echo $dbId?>">Delete</a></td>
+                        <td><a href="single_data_edit.php?edit_Id=<?php echo $dbId ?>;">Edit</a> || <a onclick="return confirm('Are You Sure?')" href="delete.php?id=<?php echo $dbId?>">Delete</a></td>
                     </tr>
                 </tbody>
